split class vs exported functions
