+++ tests/fm_dev/rti1516eLv/CMakeLists.txt
@@ -17,9 +17,10 @@ option(BUILD_TEST_APP "Build test application" ON)
 
 # variables for source and headers
 set (LIB_RTI1516ELV_SOURCES
+    rti1516eLv.cpp
+    rti1516eLvFederate.cpp
     rti1516eLvError.cpp
     rti1516eLvHelpers.cpp
-    rti1516eLv.cpp
 )
 
 if(BUILD_LIBRARY)
