+++ tests/fm_dev/rti1516eLv/rti1516eLv_errors.h
@@ -5,21 +5,6 @@
 // extern "C" {
 // #endif
 
-//#include "rti1516eLv_EXPORTS.h"
-// #include <Options.h>
-// #include <StringUtils.h>
-
-// RTI specific headers
-// #include <RTI/RTIambassadorFactory.h>
-// #include <RTI/RTIambassador.h>
-// #include <RTI/NullFederateAmbassador.h>
-// #include <RTI/time/HLAinteger64TimeFactory.h>
-// #include <RTI/time/HLAinteger64Time.h>
-// #include <RTI/time/HLAinteger64Interval.h>
-// #include <RTI/encoding/BasicDataElements.h>
-
-// #include "cintools/extcode.h"
-
 // using namespace rti1516e;
 #define LV_ERROR_RANGE_OFFSET   -500000
 #define LV_WARNING_RANGE_OFFSET 500000
